extract useWrapLines into own file
