+++ packages/react/src/recipes/seo/Article/Article.tsx
@@ -1,6 +1,6 @@
 import {ChevronDownIcon, ChevronUpIcon, MoonIcon, SunIcon} from '@primer/octicons-react'
 import clsx from 'clsx'
-import React, {useRef, useCallback, useEffect, useState, Fragment} from 'react'
+import React, {useCallback, useEffect, useState, Fragment} from 'react'
 import {
   AnimationProvider,
   Box,
@@ -28,61 +28,7 @@ import placeholderImage from '../../../fixtures/images/background-poster-ai.png'
 
 import styles from './Article.module.css'
 import {Themes, themeDetailsMap} from '../helpers'
-
-const useWrapLines = (text?: string) => {
-  const ref = useRef<HTMLHeadingElement>(null)
-  const [lines, setLines] = useState<string[]>()
-
-  useEffect(() => {
-    if (!ref.current || !text) return
-
-    const wrapLines = () => {
-      const element = ref.current
-
-      if (!element || !text) return
-
-      const originalColor = window.getComputedStyle(element).color
-
-      element.style.color = 'transparent'
-      element.textContent = ''
-
-      const words = text.split(' ')
-
-      const lineBreakPositions: number[] = []
-      let prevHeight = 0
-      const linesArr: string[][] = []
-
-      for (const word of words) {
-        element.textContent += `${word} `
-        const height = element.offsetHeight
-
-        if (linesArr.length === 0) {
-          prevHeight = height
-          linesArr.push([word])
-          continue
-        }
-
-        if (height > prevHeight) {
-          lineBreakPositions.push(linesArr.length)
-          linesArr.push([word])
-        } else {
-          linesArr.at(-1)?.push(word)
-        }
-
-        prevHeight = height
-      }
-
-      element.style.color = originalColor
-      element.textContent = ''
-
-      setLines(linesArr.map(line => line.join(' ')))
-    }
-
-    wrapLines()
-  }, [text])
-
-  return {ref, lines}
-}
+import {useWrapLines} from './useWrapLines'
 
 /**
  * This is an example of a remote content that can be fetched from a CMS or a markdown file,
